// src/lib/sites/types.ts
export type NewSiteInsertInput = {
  name?: string | null
  address?: string | null
  site_manager?: string | null
  site_assistant?: string | null
  client_manager?: string | null
  client_phone?: string | null
  start_date?: string | null
  end_date?: string | null
  quote_date?: string | null
  construction_start_date?: string | null
  inflow_path?: string | null
  work_kind?: string | null
  status?: string | null
  contract_type?: string | null
  budget?: number | null
  memo?: string | null
}